Fix parent of extra control not being a wxFileDialogBase on macOS
A wxEVT_UPDATE_UI handler expects the parent of the extra control panel
to be at least a wxFileDialogBase. However with wxMac when showing file
types the extra control gets reparented and has the panel containing
file type controls as a parent.

Fix by not reparenting in case the extra control itself already is
a panel, and then instead use the extra control as a container for
the file type controls.