Remove hardcoded max width from STC autocomplete listbox
The Scintilla engine will automatically truncate lines in the
autocomplete listbox based on the call to AutoCompSetMaxWidth().
It is not good to do it in the wx implementation code as well,
since then the meaning of AutoCompSetMaxWidth(0) (e.g. no truncation)
is not valid and the list items will be truncated anyway.

Closes https://github.com/wxWidgets/wxWidgets/pull/2250