Add a note about using ranged-based for loop with wxString
It is surprising that "for (auto& c: s)" doesn't compile, so document
this and propose a replacement.

Do not document the fact that "for (auto c: s)" actually allows to
modify the string via "c" (which is still wxUniCharRef) as this is
probably just going to confuse people more than help.